Move classmethods to tests - they are not meant to be used by users
